Deprecate ToolkitTask, add functionality to PromptTask
