+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rulesets can now be serialized and deserialized.
 - `ToolkitTask` now serializes its `tools` field.
 - `PromptTask.prompt_driver` is now serialized.
+- `PromptTask` can now do everything a `ToolkitTask` can do.
 
 ### Fixed
 
@@ -31,6 +32,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `StructureVisualizer.query_params` for setting query parameters on the url generated by `StructureVisualizer.to_url()`.
 - Parsing `ActionCallDeltaMessageContent`s with empty string `partial_input`s.
 - `Stream` util not properly propagating thread contextvars.
+
+### Deprecated
+
+- `ToolkitTask`. `PromptTask` is a drop-in replacement.
 
 ## [1.0.0] - 2024-12-09
