+++ griptape/tasks/__init__.py
@@ -1,7 +1,7 @@
 from .base_task import BaseTask
 from .base_text_input_task import BaseTextInputTask
-from .prompt_task import PromptTask
 from .actions_subtask import ActionsSubtask
+from .prompt_task import PromptTask
 from .toolkit_task import ToolkitTask
 from .text_summary_task import TextSummaryTask
 from .tool_task import ToolTask
