+++ griptape/templates/tasks/prompt_task/assistant_subtask.j2
@@ -1,0 +1,4 @@
+{% if subtask.thought %}
+Thought: {{ subtask.thought }}
+{% endif %}
+Actions: {{ subtask.actions_to_json() }}
