+++ griptape/templates/tasks/prompt_task/user_subtask.j2
@@ -1,0 +1,5 @@
+{% if subtask.output %}
+{{ stop_sequence }}: {{ subtask.output.to_text() }}
+{% else %}
+Please, keep going!
+{% endif %}
